SDK-1459. Fix infinite loop when receiving notifications of changes inside excluded folders.
Also fixes icon display in windows explorer for items inside excluded folders

This case was always a bit broken, with the DIREVENT being requeued, but with an adjusted LocalNode* and relative path that didn't match anymore.  The LocalPath changes introduced another bug that meant that the requeued event got requeued again indefinitely in this case.

The fix still requeues the event, but with the correct and matching LocalNode* and relative path.  The main fix for this case though is that the check for excluded paths is moved before that requeue check, so for this case we don't even reach the requeue code anymore.

Also the check for excluded paths is updated to check if a full path is contained within and excluded path.  This also fixes the case of sync icons showing inside excluded folders in Windows Explorer.