Avoid too big a gap between the earliest and latest piece of an upload.
If the gap gets too big, the server may return -4.
This can occur if, for example, one upload connection runs slow for some
reason (slow disk, network issues for that connection)
Limit the gap to 256MB for now, connections are paused before actual
post() until the earliest connection completes, in that case.