Avoid storage overquota when foreign
If the target folder of a putnodes is foreign (ie. an incoming shared
folder) and the API returns -17, it's not the own account which is
overquota, but the sharee's account --> do not enter into overquota
state, simply fail the operation with such error. The app can check if
the target was private/foreign to adjust the error message.

In case the putnodes is related to a sync, we should disable the
corresponding sync.