Add handling for duplicate sync configs.
This should only result if the program is terminated before the client
receives a result from the SDK telling it that the migration of a config
has a completed.

The conflict resolution strategy is simple: If two sync configs exist
for the same root, the newest is the one that is stored.