build_from_scratch: skip include file checks on windows
If the headers are not present we will fail anyway.  And cmake was outputting some errors when trying to run those checks on windows.

Also, swapping vcpkg commits in vcpkg_clone wasn't working as now there are some files/folders that get left behind and it won't overwrite them by default, use --force to make it happen