Solve some rename-and-replace cases in the reworked sync algorithm
syncItem() is split into two now to separate the move detection and do that first for all nodes in a folder.
Also sped up scanning - allow syncs to queue another scan request if that sync's last request completed already.
No need for scanTargetReachable anymore as we won't get stuck on a single unvisited scan anymore
per-node scan requests are put in the rare fields.
Implemented rare field clearing
Also moved ScanService to filesystem.h, as we no longer need the matches() test, and we can make it independent of LocalNode.
Had to rearrange include dependencies a little as it needed FSNode and filefingerprint, which in turn needed InputStreamAccess etc