SDK-1530. Merging minor improvements from sync rework branch, to avoid too many differences building up
Sync callbacks via MegaApp pass LocalPaths instead of LocalNode (only the path from LocalNode was used, often nothing from LocalNode was used)
Merged the attr_map class which captures attribute changes to be applied to Nodes.  In this version, we apply the changes before sending the request to servers (ie, SIC).  Sync rework will change that so that attribute changes are only applied to the Node via the resulting actionpackets.  But in the meantime, we still update with Speculative Instant Completion, but using the same technique (attr_map) to specify the change.
Use unique_ptr for sctable
Some clarified names
Improved debug logging for megacli and integration tests (time specified to millisecond)
Some additional commands in megacli to assist testing, especially for syncs