Fix error handling for incorrect JSON
If the JSON is invalid, it will fall to the `default` and it will
eventually fail to consume the JSON objects. In that case, the code must
return, and return `false`.
On top of it, if there was a transfer associated to this command and
it's canceled, the command must not call the completion function, since
it may lead to access to invalid memory (data from the canceled
transfer, which might have been destroyed already).