Removed 'name' member of LocalNode
It should have matched the synced Node's name but didn't always - usually it had been set to toName() of the localname member.   We can calculate that anytime, so no need to store it and have that ambiguity.

Made some other small related tidy ups, removing unused members and commented out things that are for sure replaced or not needed.

Removed cloudRoot() from Sync, which used Node* which could invite threading issues.  Now we use a CloudNode snapshot

commented a fair few assert()s that were supposed to be making sure the name field was correct (or within tolerance)