Rework sync triplet matching so we combine all three elements in a single sort now.
Some of the auto tests for unescaping were failing, and made it apparent we needed something a bit clearer for that, and able to flexibly match on escapes across all three elements of a triplet.
Adjust comparison of escapes in utfCompare, so that we consume an escape sequence if we can, but otherwise we can still match on the unescaped character (makes a difference for escaped sequences in cloud).
Add Node::latestFileVersion()
Clarify whether we are looking for the exact node or the latest file version when calling lookupCloudNode