use NodeHandle as the key for the nodes map.
this is a fix for the sometimes failing SyncPersistence test.
The failure can really only be explained by failure to look up and node `handle` (8 byte value) after creating a folder and receiving that node handle.
Likely, we still somewhere have the problem that the upper two bytes of the 8 byte handle are not zeroed.   We can avoid this issue by using NodeHandle which takes care of the 6 byte aspect.