Update some command completion functions to have NodeHandle,Error arguments
This makes them compatible for easy chaining as occurs on the sync rework branch
Also included a big performance boost for the case of 300k transfer list, and the transfers completing are all at the front (or end) of the list.  Found/fixed on sync rework.
Updated some other command arguments to NodeHandle where they were just handle.
Also included a check that we are not modifying databases from multiple threads (via transaction committer class)