SDK-1605. Add UploadHandle and NodeOrUploadHandle classes to clarify what his happening during the upload of media files
There are a lot of combinations and just using `handle` for everything is quite confusing.
 - generating thumbnail/preview for a Node or for an upload in progress
 - the various steps: generation on a separate thread, attach to existing node, attach to upload in progress
 - possible failures at each stage and whether we need to call the app back, if it was an app request

Anyway, there are quite a few data structures and it's immensely clearer if the C++ static type system keeps track of whether
we are dealing with a Node or an Upload, by what type of handle is in the data structure.

This addresses the recently discovered issue where an UploadHandle (as a handle) was passed to nodebyhandle().  (discovered thanks to an assert on what is a valid NodeHandle or not)

Additionally this commit contains some comments where it's now clear that we have some gaps in the error processing, where for example generation for Uploads was considered, but not generation for existing nodes.