Return null if key is invalid
If the key is invalid, the `setkey()` returns false and the cipher is
not ready. Hence, it should return nullptr instead of the (potentially
invalid) cipher.
Note the recycled cipher is used once and again with different keys, and
previously used keys might have become invalid by the time the cipher is
reused, which would result in crashes in Crypto++.