Fix intermittently failing tests BasicSync_MoveLocalFolderPlain, BasicSync_MoveLocalFolderBetweenSyncs.
Ensure we wait for the first actionpacket to arrive before the general waitonsyncs()
In this particular case, client2 did not receive the actionpackets for the move, even though client1 had, within the 8 second window that waitonsyncs() would wait itself.
Now we give it up to 30 seconds for those first indicators, but we don't have to wait more than before for the normal successful cases.