Fix ReplaceParentWithEmptyChild test.
The issue that this test hit was that we only allow a move to proceed if
the source and target have a well-defined exclusion state.

Also note that generally we don't descend below move targets.

The reason this test hit this issue is that the source is below the
target and since we never descend to the target, we can never compute
its exclusion state. The result is a stall.

The solution is to allow the engine to descend to the source, acting as
if it is recursing below a removed node. This let us compute the
source's exclusion state which will allow the move to complete.