SDK-1651. Bring minor changes from sync rework (6th stage)
Changed completion function for addsync and some others to avoid any pointers to sync-thread managed data structures (they will be on a thread in sync rework) - now the callback fields are all by value.
Clarified the difference between the State field for `sp` and `sphb`, make sure we don't use the enum for one in the values for the other.
Update checkSyncConfig to avoid supplying a Node* back to the caller (ie sync thread sometimes) which only needs the node name.
startSync() to take a node handle instead of Node* for the same reason.
Moved easily detectable sync errors to checkSyncConfig, things that can be checked before attempting to actually starting the sync (so we don't need to call back from the sync thread).
megacli:  added sync rename command