Perform the macsmac calculation as we go for the contiguous-from-start portion of the macs map, during uploads and downloads
This lets us keep the mac map quite small, as we can collapse the leading entries into a single entry with the result so far.
For extremely large files, not doing this causes uploads and downloads to slow down and spend a lot of time saving large mac maps to the database frequently.
We have to be careful to not compromise the fix for the historic files that had slightly bad macs due to missing out one or two chunks from the calc on upload.
Quite a lot of code has been factored into the chunkmac_map, that was previously scattered about.  This result is much more modular.