Report scan-blocked paths by the stall mechanism
Found a much simpler way to do this (that doesn't flip-flop the state when we retry the scan).
Also fixed an issue with deleteq - it could contain deleted LocalNodes.   Now we use the delayed-committer system and delete the database record straight away.
Ran into some issues when combining scan-blocked and then editing/renaming .megaignore to bring it in and out of ignore state.
Fixed those mainly by adjusting the exclusion filter, scan-blocked are TYPE_UNKNOWN (at least initially) so we should consider both file/folder rules for that case.