LocalPath now always contains absolute paths, simplifications from local2path, path2local moved into LocalPath.
This enables LocalPath, for windows, to always be prefixed with \\?\ which enables accessing files with long names, long paths, trailing spaces or '.', etc.
LocalPath transparently adds that prefix internally, and we don't need the large number of #ifdefs and special code adding that inconsistently, LocalPath takes care of it all.
Constructing a LocalPath from a string (for windows) converts it internally, the function do to so specifies whether it is a relative or (should be an) absolute path.
For those cases where a relative path is passed where ideally the caller would provide an absolute path, we conver it.
The process to do that for windows is much improved (and factored into one spot in LocalPath).
Moved local2path and path2local into LocalPath, the main (and nearly only) user of those.  This means fsaccess is not needed to be passed around nearly so much.  Many or most call sites affected were already affected by the above change.
LocalPath::platformEncoded is now almost not used except internally - reducing this was a long term goal
Removed some functions from the intermediate layer relating to that and syncs, which should not be used by clients, at least in sync rework.
Removed FileSystemAccess::getPathSeparator, hardly used anyway, replaced by simple LocalPath::localPathSeparator_utf8 character
Removed LocalPath::ensureWinExtendedPathLenPrefix, now built in