syncs_restored(...) now accepts a SyncError parameter.
The rationale behind adding this parameter is that we may actually fail
to resume syncs on startup before we attempt to resume any specific
sync.

An example of how this could occur if if we are unable to create a
SyncConfigStore or if we're unable to read the internal sync
configuration database.

In either case, we're unable to load any configs from disk so we're
unable to report an error via sync_auto_resume_result(...).