Adjustments for dealing with local move/rename while the initial upload is occurring.
We remember the parent node handle and target name of uploads, and recognise them that way when processing newly appeared cloud nodes. And then set the resulting LocalNode up so it'll be recognized as needing to be moved.    And right now, from manual testing, that seems to be working - lots of tricky issues overcome to get to this point, it's a significant refinement on the new algorithm.