Resolved PUT file versioning cases for sync and manual uploads in File::completed
With versioning off, the old file is moved to cloud sync debris
with versioning on, the old file becomes the first version of the new file.
For manual uploads, we still send ov but let the API decide how to interpret it depending on the global file versioning flag.
It may be permanently deleted (but retain the old files' versions) or it may become the first version of the new file.
Converted ovhandle to NodeHandle