You can now upload duplicate files with uploadFile(...).
To achieve this, the flag fixNameConflicts has been added to File.

The purpose of this flag is to signal whether we should ignore the
usual rules in effect deciding what we should do when we manually upload
a file that already exists.

If this flag is false, we'll always create a node in the cloud.