Avoid overwrite th in putnodes command if it's not accesible
While a file is beign uploaded, destination target could be removed, or
in case of inshare, permissions could be revoked by sharer.

In this case target handle will not be modified for putnodes (preserve
original one), is API responsibility to check if target is accesible by
user, and if it's not posible, API will put file into Rubish bin.