Added syncdebug_notification(...) callback to MegaApp.
This callback is intended only for debug and/or testing use.

The purpose of this callback is to give clients the ability to perform
some extra processing after a notification has been checked for validity
but before that notification has been processed.

An example of where this could be useful is when we're testing how the
sync engine behaves with weird fingerprints. Say, a file's content has
changed but it's fingerprint has the same mtime and size.

We need the mtime (or size) to be different in order for the
notification to not be skipped but we want them to be the same when the
notification is actually processed.