Ensure client up to date
Since the test creates a new node and immediately performs a
locallogout, the DB transaction with the new node might be rolled back.
In result, upon resumption, the test checks the existence of the new
node and it might not be received from API yet (in an actionpacket).
By waiting for the `NODES_CURRENT` event, it is ensured.