Reset session id
If kept the same after a `locallogout()`, the DB transaction can be
rolled-back and the missing updates (via API commands with `"i"`) will
be ignored when received upon session's resumption, as action packets.

The session id is actually a client's session, not user's session.
Hence, it should be reset to avoid the aforementioned issue.