Many changes to support cross-thread (client/sync) distribution of downloaded files to their final paths, and support efficiency in transfer callbacks in MEGAsync.
QTMegaListener now forwards the created Qt events to a supplied Qt object, and doesn't try to customeEvent() them itself.  So MEGAsync can group, optimize etc.
FileDistributor class introduced.   It factors most of the logic around move/copy of a downloaded file to the various final locations, safely across client/sync threads.
Removed "looprequested" flag.  This was causing excessively long mutex locks as multiple transfers would cause it, and the next loop with then add more, and set the flag again.
Allow a second in-progress scan for each Sync, for folders not yet scanned.  This moderates the effect of downloads completing and triggering folder rescans while we have not yet initially scanned the entire sync.
Hoist tmpnamelocal() out of the filesystem derived classes, so it can be used directly as a static function, no filesystemaccess pointer needed.
SyncDownload_inClient class now doesn't need to cause a rename of the downloaded tmp file to another tmp flie before it's available to be distributed.
SyncDownload_inClient now supplies accurate final filenames to be shown in MEGAsync transfers in the GUI (not the temp file name anymore).