Parents now track a synthesized FSNode when a child completes downloading.
The rationale behind this is that there can sometimes be a delay between
when a filesystem change occurs and when a notification for that change
is sent to us by the system.

If we don't track, in some way, that the downloaded file exists, the
engine would continue to try and download that file until it received a
notification from the system.