Remove targets from a file distributor when appropriate.
When a FileDistributor is destroyed, we check that all of its targets
have been distributed to. Unfortunately, prior to this change, this
would trip an assertion whenever we destroyed a transfer that downloaded
successfully but could not be copied (or moved) into place.

This changeset addresses this issue by adding a new flag,
wasDistributed, which is set when the engine successfully copies (moves)
a download into place.

When a SyncDownload_inClient instance is destroyed, we check whether a
file was successfully moved into place (distributed.) If we couldn't
move the file into place, we remove one target from the distributor.