Solve REMOTE_PATH_HAS_CHANGED issue for tests.
The test were overcomplicated because the "old path" was getting adjusted every time the node moved, even if the sync wasn't running.  So really it was the "current path".  To know that, you can just look it up from the handle, no need to even store a path.
And tests were waiting for fail state after making a second adjustment, but of course the sync had already failed.
Anyway, adjusted it so that the condition is detected once and fails the sync.   Then it'll stay in that state (remembering the original path) until the user starts the sync, at which time the "old path" is reset.
Other minor build fixes etc.