Fix wait time indication for out-of-transfer-quota pro users
Since out-of-transfer quota error (509) may happen for streaming and
non-streaming transfers, the calculation of the back off duration has
been isolated (in Utils::overTransferQuotaBackoff method) instead of
duplicating the logic.

Defined business logic is that any pro user who runs out of transfer
quota shall wait for a renewal or an upgrade.

The solution implemented adds a member variable -mMyAccount- (and
struct -MyAccountData) to MegaClient for the sake of avoiding constant
queries to the server. Instead, the pro level, and time until the user
keeps the pro level are saved in memory.

These values (pro level and time until pro level status) are refreshed
with every "uq" command which includes "utype" (for pro level) and
"suntil" (for epoch time until the pro level status is valid).

In addition, a new "uq" command is requested as the final step of a
successful login; and, the member variable is reset at the end of a
successful logout.

Note as well that event 99408 is meant only for non-pro users when the
X-MEGA-Time-Left respose header is negative.