Fix the login memory leak due to the new AccountDetails object
In order to do so, MegaRequestPrivate::accountDetails is refactored
from AccountDetails* to a std::shared_ptr<AccountDetails>.

Once the refactor is completed, the memory leak is handled by the
smart pointer created in CommandLogin::procresult call.

As part of the refactor, there are some methods which just consume the
data pointed by the std::shared_ptr, such as:
* MegaApp::account_details(AccountData*, error)
* MegaApp::account_details(AccountData*, bool, bool, bool, bool, bool,
bool)
* MegaAccountDetailsPrivate::fromAccountDetails(AccountDetails*)

Instead of replacing these methods, the raw pointer from the
std::shared_ptr is passed as a parameter for the sake of keeping the
refactor as small as possible.

In case any client implemeting MegaApp is expecting ownership of the
pointed memory, there will be an error when accessing since the
std::shared_ptr will keep ownership.