Local ignore files remain in effect if there's a remote name clash.
If there are two (or more) local ignore files then signal a name clash
and wait for the situation to be resolved.

If there is no local ignore file present and there are two (or more)
ignore files present in the cloud then signal a name clash and wait for
the situation to be resolved.

If there is a single ignore file present and there are two (or more)
ignore files present in the cloud then the local ignore file remains in
effect. If the local ignore file is later removed then we revert to the
case directly above.