SDK-1946. Merging minor changes from sync rework, 7th stage
Fixes for networking on windows: make sure to call curl immediately when a socket changes from read to write mode, and prevent the curl call from being skipped due to flag `curlsocketsprocessed`.
Assert added to ensure we always do call curl after we clear the OS event that notified us that we should.
Introduce proper types for Uploadtoken and FileKey, with sub-structure
Introduce `action` as a member of MegaRequestPrivate.  We can use it to radically simplify the code layout for intermediate layer functions.
Use NodeHandle for NewNode's nodehandle.
Finish off the addition of a completion function to all forms of putnodes()
Convert getovhandle() to getovnode(), sometimes we need the node, and the handle can be retrieved from the node.
rename LOCAL_FINGERPRINT_MISMATCH to LOCAL_FILESYSTEM_MISMATCH, as the old name was misleading - it's not a FileFingerprint.
Other small minor function additions etc that can easily be brought back with no risk