postLog() should go via SimpleLogger so it forwards to the correct output mechanism
Moved postLog from externalLogger to SimpleLogger
Removed =false for singleExclusiveLogger flag for addLoggerObject and removeLoggerObject, so the compiler reminds us to check each app and supply the right value for that app's log usage.
setLogLevel() also forwarded direct to SimpleLogger, no need to go via ExternalLogger, especially if exclusiveLogger is the one in use.
Confirmed postLog() is working in integration tests.