Improve notifications upon new versions of shared nodes. Debugging
Current implementation is still WIP and includes debugging traces that
will be removed in final version.

Current implementation covers only the portion of supressing the
multiple notifications for new files/version for each version. It
still misses handling the supression of file removal messages for the
node deletion on the previous versions.

Changes in this commit involve:
- creating the new alert update type for the shared node
- logic to modify alerts in the sense of detecting that older versions
are not new nodes
- said logic (e.g. `isHandleInAlertsAsRemoved`, or
`setNewNodeAlertForHandleToUpdate`)has been placed in-lined as part of
development cycle. It is not its final design location
- debugging traces to locate if -and which- `notifypurge` is called
between the deletion of old versions/nodes and the readdition