Protect against removed remote node
It should never happen that the remote node of a backup is already
removed at this codepath (assert will fail). However, in case there were
a bug and the node were removed without removing the backup, we want to
protect the code: do not attempt to delete/move the node and show a
warning in log.