The sync engine is now responsible for checking if an upload results in a filename anomaly.
The rationale for this change is that where we're uploading to might
change while the upload is in progress. If the client is responsible for
checking for an anomaly, it'll be checking against an outdated name.

Making the engine responsible allows us to check against the latest name
without introducing any races.