Review: Prevent repeated scans when under heavy load.
The concern here was the system might issue several "recursive"
notifications for some path before a "recursive" notification was
processed by the engine for some parent.

Our solution here is to make use of the "subtree rescan" flag already
present in the engine. When the engine detects a recursive notification,
it tries to locate the exact node corresponding to the notified path. If
that node is found, mark it and all nodes below it as requiring a
rescan.