SDK-2057. Make sync stall objects easier to interpret for GUI display
Provide two cloud paths and two local paths, always (empty if the path is not relevant for the case)
Provide an enum code with each path to indicate to the user what the issue with that path is (0 for paths that don't need a message)
Consolidate the set of enum codes for overall stall cases, since the path problems give a suitable sub-code.
Generic-ize the intermediate layer methods for retrieving said paths and the associated path problems.
That generic interface might mean we can supply name clashes the same way.