Reduced the number of distinct test function specializations.
This changeset introduces the CloudItem class which exists to represent
a specific node in many ways. It can be created implicitly from many
types such as a path, a node reference or a node handle. Instances of
the CloudItem class can be "resolved" to yield a pointer to a node, if
available.

There are two main reasons for this class. First, we have many, many
distinct specializations for various operations such as moving a node.
Many of these specializations were added by me for convenience.
Maintaining all of these specializations takes time and makes it harder
for newcomers to figure out how to use our code.

The second reason is that we may use different things to denote a remote
node at different times. Sometimes we'll have a node reference handy and
others, we'll have only the path. Sometimes, we might want to specify
one parameter by handle and another by path. The CloudItem class makes
this possible.