Fix an extremely unlikely but possible issue where there could be a missing chunkmac in the final mac-of-macs for an upload
how it occurred before:

the original upload attempt stops when a chunk A has been sent, the server accepted it, but the program exited before receiving the result, and it left a gap in the MACs stored in the DB for that transfer.

the transfer is resumed on restart, chunks start encrypting and uploading.  But the last chunk is small and uploads before encryption of  the chunk A is complete.

On processing the uploadtoken, we assume all the other chunk macs are present and calculate the mac-of-macs, but chunk A's is missing.

A new command is added in megacli to hammer this case for testing:  cycleuploaddownload