Fix: Make sure selected tests wait for stuff to happen.
The fingerprint-collision and move-target-has-filesystem-watch tests
have recently failed on Jenkins. Digging into the issue, it appears that
the problem is late action packets. That or variation in timing due to
say, the upload delay.

This changeset attempts to resolve this issue by forcing these tests to
wait until changes have hit the cloud using a wait-predicate.