Add tests to confirm the API for file upload with custom mtime
Also fix the case so that the tests work
Made the ability to specify an overriding mtime explicit in sendPutnodes (the secrecy of the old technique is how this bug got in)
Add a few comments for the override field, make it clear that MegaApi::INVALID_CUSTOM_MOD_TIME signals the not overridden case.
Custom completed() override for MegaFilePut that passes the overridden mtime if appropriate
Fix the cases when we copy an existing node or think that the uploaded file is already present, for those to happen they must have the overridden mtime already.
Noticed that string::data() is used a lot when c_str() is more appropriate (and correct, for pre-c++11) for adapting a std::string to be passed as a c string (it's right there in the name!) so took the opportunity to update those in SdkTest_test.cpp, and set a better example for new programmers.