Fix the original problem
THe original problem was that we tried to fix in commit:
8cd0f144384f2ef6d2b295f6e47b7ccecbe93625 was that
assert(transferMap.empty()) failed in some cases.

In original code, folder transfers were skipped to be removed at
abortPendingActions, as it was expected that it's last subtransfer will
remove it.

In case we have a folder transfer without any subtransfer that trasnfer
won't be removed nor informed to the app and also
assert(transferMap.empty()) will fail.

We just need to leave original code (with some improvements) and also
remove inmediately those folder trransfers without subtransfers.

Currently we was skipping folder transfers removal at abortPending
(delegating this task to subtransfers) so