Make sure we complete() the upload/download folder transfers once exactly for all cases
Fix the case where we try to put all the transfers on the queue atomically, but the cancel token could interrupt and leave a portion yet to be processed

Cancel token no longer needed for sendPendingTransfers due to that
When the token is activated while that is going on, the sub-transfers all refer to the same token.  Each of those is extremely quickly processed, to the remainder of the batch.

Avoid logging too much when cancelling extremely large batches of transfers.
Fixed a case where we were trying to determine filesystem type too frequently.
After that, very large sets of transfers (500k) were processed in just a few seconds

Rearranged the upload controller and download controller sequence so we always call complete() exactly once.
Previoulsy the logic was unclear, with hasEnded() possibly executing that out of line, possibly more than once.

For uploads, de-duplicated the code for actual queueing of the file transfers

Took care not to use the controller after they may have completed themselves, and therefore also deleted themselves.

Notification of STAGE_TRANSFERRING_FILES moved to the point where the first sub-transfer is started, avoids that use-of-deleted case

also avoided some excessive error logging for pre-existing favourites with attribute 0