SDK-2195. Possible failure to call onTransferFinished after the final sub-transfer of a folder transfer finishes
Add logging to confirm that we finish folder transfers properly, and call onTransferFinished for the folder transfer.  Also count down as subtransfers finish.

Adjust genUploadTransfersForFiles to check for cancellation and return false if it was cancelled (brings it in line with genDownloadTransfersForFiles)
Adjust sendPendingOperation to speed up the case of cancelling a folder transfer, if there are huge numbers of sub-transfers.  Turns out it can take quite a while to start/stop all remaining transfers, with many callbacks etc.

For upload folders, gather the FileFingerprint for each file while we are scanning.  This is the slow part, and should happen during scanning when we're not taking time while the main mutex is locked
Added an extra parameter to createUploadTransfer so that it can be supplied with a precalculated fingerprint
Updated createUploadTransfer to calculate the fingerprint at that time if not supplied.  This keeps that time consuming operation outside the main mutex lock, for manual individual transfers.

Factor MegaFolderUploadController and MegaFolderDownloadController's onTransferXXX function to their base class MegaRecursiveOperation as they were exactly identical anyway
Also their complete() functions, also identical

Remove unused function completeRecursiveOperation

MegaTransfer object records the pre-calculated fingerprint for upload files now