Remove wrong condition
tlv->get(h, buffer) extracts alias from TLVstore, and it seems that
removed condition expect that it's length is NODEHANDLE length
(probably tought that it was comparing against user handle).

anyway we don't need to check length just search in TLVstore, if it's
not found request will fail with an error.