Relax restrictions to create device/backup folders
Upon creation of new backups, if the local folder is the same than a
previously configured backup in the current device, it will create a
duplicate.

Also, if the backup requires to create a new device folder, and there's
another device folder already under "My backups", it will create a
duplicate.