Some fixes and improvements
1) Fix socket disconnection so it starts again from REQ_READY (otherwise
it will be REQ_FAILURE after calling disconnect(), and that will cause a
reset of the transfer)

2) Improve variable names, calculation and logic for
StreamingBuffer::init() so everything is padded to MIN_DELIVERY_CHUNK
(unless the requested capacity is a smaller value which doesn't need to
be truncated)

3) Set HTTPCTX maxBufferSize and maxOutputSize values after the new
calculations done by init() so both of them are coherent.

4) Add reset() to "forget" the space used in the streaming buffer by the
prior headers-write operation, so it doesn't affect the
MIN_DELIVERY_CHUNK alignement.