Prevent already uploaded chunks to be reuploaded
This npos re-calculation was not taking into account the already
processed chunks. Whereas it is not a problem regarding transfer
integrity or upload completion (even if they were already uploaded
they're still correct chunks, so it should not be a problem), it
affected transfer completed progress counter (affecting the apps at
least in a "cosmetic" way), apart from the extra -small but
unneeded-upload effort.