Adjustments to dealing with duplicate LocalNodes in the sync database on start.
The earlier one from the database is removed, the later one kept (same as the one that ends up being used, in the old system).
Children maps are simplified with normal keys for the key-value pair to avoid mistakes
Updates of the children maps make sure we only remove the child if it really is us, from the LocalNode destructor.
Update some cases where DBTableTransactionCommitter should be used, for non-transfer databases.