Add MegaUserAlert::getPcrHandle()
When IncomingPendingContact is created from action packets, it gets its
handle set to the UserAlert::Base::userHandle. However, when it's
created from `sc50`, the user handle is never set.
This commit ensures the user's handle is set in both cases (for
backwards compatibility). Additionally, it introduces a dedicated member
to store the PCR's handle, and also a public interface to use for the
related alerts.