Make the callers of LocalPath::toPath() specify whether the string should be utf8-normalized or not (usually not).
Previously the philosophy was to normalize all strings everywhere for easy comparison.  However, then we can't tell exactly what is going on in the logging when utf8 strings fail to match, but the logs show they are exactly the same.  Also, filesystems can be sensitive to the encoding - we can put two files into NTFS with different normalizations of the same string, and they are separate files with separate names.   Therefore, we do need to be precise and not normalize filesystem-side strings.

This may also help with file paths passed to 3rd party libraries.  If we changed the normalization, perhaps the library won't find the file, or find the wrong one.