Merge canChangeVault flag from develop/backupRework
Merge one small aspect of develop, since the backup rework changes are there now and are too big and too complicated to merge in one go.
This one is a manual copy over of one small aspect of that, plus a few bug fixes along the way (of the incoming changes).