SDK-2350.  minor changes from sync rework (18th stage) - LocalPath::toPath optional normalization
Bringing more changes from sync rework to try to make develop and SRW more compatible after BRW was merged to develop.

toPath used to always utf-normalize (to one particular normalization), without option. However, that was a problem in sync rework where there were some tricky cases and we could not tell what was going on in the logs, because the path names output in the log were all normalized the same way, so we couldn't see the actual ut8 that was causing the problems (with comparisons between local and remote names).

Add a flat to toPath() so that normalization is optional, and don't set it when logging local paths. In existing usage, pass false if the path is all-local. Pass true if unsure or the path is compared to Nodes, or for backward compatibility.