SDK-2355. minor changes from sync rework (19th stage)
Trying to reconcile Backup Rework changes that are on develop, with Sync Rework.

Comparing diffs, looking for chances to make them more similar

So far, just very minor things, but ones that help by reducing the volume of changed lines in the diffs

Including some functions for managing cross-thread operations, that will be come relevant in sync rework, when we really do have a separate thread for syncs.
In the meantime, these are just stubs that call the operation immediately, all on the same thread.