Manual management of the HTTP redirections
Using CURLOPT_FOLLOWLOCATION had two undesired effects. On the one hand,
cURL manages DNS resolution by itself (probably blocking the thread). On
the other hand, CURLOPT_PINNEDPUBLICKEY can't be modified for the second
request. In this case, the public key of reqstat.api.mega.co.nz (the
target of the redirection) is the same as for g.api.mega.co.nz and cURL
seems to reuse the public key for the second request. But that behavior
is not properly documented and could change in the future.

I have also added the flag expectredirect for backwards compatibility,
because until now a HTTP 3xx response code was considered an error in
the request, so I think that it's good to keep that behavior for
requests not expecting those responses.

Since I have modified code in the function CurlHttpIO::multidoio of
net.cpp, I have also added req->logname in logs generated in that
function that didn't show it.